refactor: 코드리뷰 반영 ( 인프라 계층에 대해서 고찰)
