+++ src/main/java/kitchenpos/eatinorders/infra/JpaOrderTableRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.eatinorders.domain;
+package kitchenpos.eatinorders.infra;
 
+import kitchenpos.eatinorders.domain.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
