+++ src/main/java/kitchenpos/eatinorders/infra/OrderTableRepository.java
@@ -1,4 +1,6 @@
-package kitchenpos.eatinorders.domain;
+package kitchenpos.eatinorders.infra;
+
+import kitchenpos.eatinorders.domain.OrderTable;
 
 import java.util.List;
 import java.util.Optional;
