+++ src/main/java/kitchenpos/menus/infra/MenuGroupRepository.java
@@ -1,4 +1,6 @@
-package kitchenpos.menus.domain;
+package kitchenpos.menus.infra;
+
+import kitchenpos.menus.domain.MenuGroup;
 
 import java.util.List;
 import java.util.Optional;
