+++ src/main/java/kitchenpos/menus/infra/MenuRepository.java
@@ -1,4 +1,6 @@
-package kitchenpos.menus.domain;
+package kitchenpos.menus.infra;
+
+import kitchenpos.menus.domain.Menu;
 
 import java.util.List;
 import java.util.Optional;
