+++ src/main/java/kitchenpos/products/common/NamePolicy.java
@@ -2,7 +2,6 @@ package kitchenpos.products.common;
 
 import kitchenpos.common.DomainService;
 import kitchenpos.products.infra.PurgomalumClient;
-import org.springframework.stereotype.Service;
 
 import java.util.Objects;
