+++ src/main/java/kitchenpos/products/infra/JpaProductRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.products.tobe.domain;
+package kitchenpos.products.infra;
 
+import kitchenpos.products.tobe.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
